moved splitting to event - tentative
